authenticate.py
from google.oauth2 import id_token
from google.auth.transport import requests
from pprint import pprint

# source:
# https://developers.google.com/identity/sign-in/web/backend-auth#verify-the-integrity-of-the-id-token

CLIENT_ID = '[CLIENTID]'
token = '[TOKEN]'

try:
    # Specify the CLIENT_ID of the app that accesses the backend:
    idinfo = id_token.verify_oauth2_token(token, requests.Request(), CLIENT_ID)

    # Or, if multiple clients access the backend server:
    # idinfo = id_token.verify_oauth2_token(token, requests.Request())
    # if idinfo['aud'] not in [CLIENT_ID_1, CLIENT_ID_2, CLIENT_ID_3]:
    #     raise ValueError('Could not verify audience.')

    if idinfo['iss'] not in ['accounts.google.com', 'https://accounts.google.com']:
        raise ValueError('Wrong issuer.')

    # If auth request is from a G Suite domain:
    # if idinfo['hd'] != GSUITE_DOMAIN_NAME:
    #     raise ValueError('Wrong hosted domain.')

    # ID token is valid. Get the user's Google Account ID from the decoded token.
    userid = idinfo['sub']
    pprint(idinfo)
except ValueError as error:
    # Invalid token
    pprint(error)
